Rename String and Base
