+++ lib/al/include/Library/Base/StringUtil.h
