+++ lib/al/src/Library/Base/HashCodeUtil.cpp
@@ -1,27 +1,27 @@
-#include "Library/Base/Base.h"
+#include "Library/Base/HashCodeUtil.h"
 
 #include <cctype>
 #include <cstdio>
 #include <cstring>
 
 namespace al {
-s32 calcHashCode(char const* hash) {
-    if (hash[0] == '\0') {
+s32 calcHashCode(char const* str) {
+    if (str[0] == '\0') {
         return 0;
     }
 
     s32 hashCode = 0;
-    for (s32 i = 0; hash[i] != '\0'; i++) {
-        hashCode = (hashCode * 0x1f) + hash[i];
+    for (s32 i = 0; str[i] != '\0'; i++) {
+        hashCode = (hashCode * 0x1f) + str[i];
     }
 
     return hashCode;
 }
 
-s32 calcHashCodeLower(char const* hash) {
+s32 calcHashCodeLower(char const* str) {
     s32 hashCode = 0;
-    for (s32 i = 0; hash[i] != '\0'; i++) {
-        hashCode = (hashCode * 0x1f) + tolower(hash[i]);
+    for (s32 i = 0; str[i] != '\0'; i++) {
+        hashCode = (hashCode * 0x1f) + tolower(str[i]);
     }
 
     return hashCode;
