+++ lib/al/src/Library/LiveActor/SubActorKeeper.cpp
@@ -1,6 +1,6 @@
 #include "Library/LiveActor/SubActorKeeper.h"
 
-#include "Library/Base/String.h"
+#include "Library/Base/StringUtil.h"
 #include "Library/LiveActor/ActorModelFunction.h"
 #include "Library/LiveActor/ActorResourceFunction.h"
 #include "Library/LiveActor/ActorSensorFunction.h"
