+++ src/Npc/AchievementInfoReader.cpp
@@ -1,6 +1,6 @@
 #include "Npc/AchievementInfoReader.h"
 
-#include "Library/Base/String.h"
+#include "Library/Base/StringUtil.h"
 #include "Library/Resource/ResourceUtil.h"
 #include "Library/Yaml/ByamlIter.h"
