+++ src/Player/PlayerCostumeInfo.cpp
@@ -1,6 +1,6 @@
 #include "Player/PlayerCostumeInfo.h"
 
-#include "Library/Base/String.h"
+#include "Library/Base/StringUtil.h"
 #include "Library/Math/MathUtil.h"
 
 PlayerHeadCostumeInfo::PlayerHeadCostumeInfo(char const* a1) {
